edit first

+++ first.js
@@ -124,30 +124,38 @@ app.post('/updateUser/:userId',multerConfig,async (req,res) => {
     const Lname = req.body.LName;
     const gender = req.body.Gender;
     const age = req.body.Age;
-    const result = await cloud.uploads(req.files[0].path);
-    const imageName =  req.files[0].originalname;
-    const url =  result.url;
+    let imageName;
+    let url;
     
+    if (req.files[0] ) {
+        const result = await cloud.uploads(req.files[0].path);
+        imageName =  req.files[0].originalname;
+        url =  result.url;
+        
+    }
 
     User.findById(userId)
     .then(users => {
         if(!users){
             
             return res.status(404).json({"message":"user not found"});
-        }
+        } else if( req.files[0] ){
+            users.imageName =  imageName;
+            users.url = url;
 
+        }
         users.Fname = Fname;
         users.Lname = Lname;
         users.gender = gender;
         users.age = age;
-        users.imageName =  imageName;
-        users.url = url;
         return users.save();
         
     })
     .then(result => {
         // delete image from local
-        fs.unlinkSync(req.files[0].path);
+        if(req.files[0]) {
+            fs.unlinkSync(req.files[0].path);
+        }
         // //delete image from local
         res.status(200).json({"message":"user updated",user:result});
         console.log(result);
